+++ app/src/main/java/com/relayhome/launcher/LauncherOverride.kt
@@ -18,8 +18,9 @@ internal data class StockLauncherOverride(
     val activityName: String,
     val label: String
 ) {
-    val disableCommand: String get() = "adb shell pm disable-user --user 0 $packageName"
-    val restoreCommand: String get() = "adb shell pm enable --user 0 $packageName"
+    val componentName: String get() = "$packageName/$activityName"
+    val disableCommand: String get() = "adb shell pm disable-user --user 0 $componentName"
+    val restoreCommand: String get() = "adb shell pm enable --user 0 $componentName"
 }
 
 internal object LauncherOverride {